requirements.txt
langchain_community
langchain
google-generativeai
streamlit
sentence-transformers
faiss-cpu